% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiscoverSL.r
\name{calculateDiffExp}
\alias{calculateDiffExp}
\title{Calculate differential expression}
\usage{
calculateDiffExp(gene1, cancer)
}
\arguments{
\item{gene1}{character. gene symbol for the primary gene. e.g. "BRCA2"}

\item{cancer}{character. cancer type code. e.g. "BRCA"}
}
\value{
A data frame with following fields: Gene2(interactor gene), Gene1(primary gene), logFC(log fold change of expression of gene2 in presence of mutation in gene1), PValue(p-value for differential expression of gene2 in presence of mutation in gene1)
}
\description{
Given the primary gene of interest and the cancer type, it calculates the differential expression of the gene list (output from findIteractor) in samples with or without mutation in the primary gene for the cancer type
}
\examples{
dg <- calculateDiffExp("BRCA2","BRCA")
}
