% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiscoverSL.r
\name{calculateDrugSensitivity}
\alias{calculateDrugSensitivity}
\title{Calculate conditional drug sensitivity}
\usage{
calculateDrugSensitivity(gene, gene2)
}
\arguments{
\item{gene2}{character vector. It can be either the output gene list from findInteractor, or a user-supplied gene list. list of gene symbols for the interactor gene. e.g. c("PARP1","ATM","NF1","EGFR")}

\item{gene1}{character. gene symbol for the primary gene. e.g. "BRCA2"}
}
\value{
A data frame with following fields: Gene2(interactor gene), Gene1(primary gene), Drug (Drug name), PvalDrug(p-value for gene2 having less RNAi score in cells where gene1 is mutated compared to cells where gene1 is not mutated)
}
\description{
Given the primary gene of interest and the interactor gene list, it calculates the difference in Drug activity score (IC50) for targeting the interactor gene in cases where the primary gene is mutated vs not mutated
}
\examples{
gene <- findInteractor()
drugscore <- calculateDrugSensitivity("BRCA2",gene)
}
