% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiscoverSL.r
\name{calculateExpCorrelation}
\alias{calculateExpCorrelation}
\title{Calculate Expression correlation}
\usage{
calculateExpCorrelation(gene1, gene2, cancer)
}
\arguments{
\item{gene1}{character. gene symbol for the primary gene. e.g. "BRCA2"}

\item{gene2}{character vector. It can be either the output gene list from findInteractor, or a user-supplied gene list. list of gene symbols for the interactor gene. e.g. c("CHEK1","ATM","NF1","EGFR")}

\item{cancer}{character. cancer type code. e.g. "BRCA"}
}
\value{
A data frame with following fields: Gene2(interactor gene), Gene1(primary gene), exp.correlation(Pearson's coefficient for expression correlation of gene2 and gene1), correlation.pvalue(p-value for expression correlation of gene2 and gene1)
}
\description{
Given the primary gene of interest and the cancer type, it calculates the expression correlation of the primary gene with the interactor genes from the supplied gene list (output from findIteractor) for the cancer type
}
\examples{
dfgene <- calculateDiffExp("BRCA2","BRCA")
gene <- findInteractor()
cg <- calculateExpCorrelation("BRCA2",gene,"BRCA")
}
