% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiscoverSL.r
\name{calculateMutex}
\alias{calculateMutex}
\title{Calculate Mutual Exclusivity}
\usage{
calculateMutex(gene1, gene2, cancer)
}
\arguments{
\item{gene1}{character. gene symbol for the primary gene. e.g. "BRCA2"}

\item{gene2}{character vector. It can be either the output gene list from findInteractor, or a user-supplied gene list. list of gene symbols for the interactor gene. e.g. c("PARP1","ATM","NF1","EGFR")}

\item{cancer}{character. cancer type code. e.g. "BRCA"}
}
\value{
A data frame with following fields: Gene2(interactor gene), Gene1(primary gene), MutexAmp(p-value for mutual exclusive amplification of gene1 and gene2), MutexDel(p-value for mutual exclusive deletion of gene1 and gene2), MutexMut(p-value for mutual exclusive mutation of gene1 and gene2)
}
\description{
Given the primary gene of interest and the cancer type, it calculates the mutual exclusivity of amplification, deletion and mutation events of the primary gene and the interactor genes from the supplied gene list (output from findIteractor) for the cancer type
}
\examples{
dfgene <- calculateDiffExp("BRCA2","BRCA")
gene <- findInteractor()
cg <- calculateExpCorrelation("BRCA2",gene,"BRCA")
gene <- as.character(cg$Gene2)
mutdata <- calculateMutex("BRCA2",gene,"BRCA")
}
