% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiscoverSL.r
\name{calculatePathway}
\alias{calculatePathway}
\title{Calculate probability of sharing pathways}
\usage{
calculatePathway(gene1, gene)
}
\arguments{
\item{gene1}{character. gene symbol for the primary gene. e.g. "BRCA2"}

\item{gene2}{character vector. It can be either the output gene list from findInteractor, or a user-supplied gene list. list of gene symbols for the interactor gene. e.g. c("PARP1","ATM","NF1","EGFR")}
}
\value{
A data frame with following fields: Gene2(interactor gene), Gene1(primary gene), PvalPathway(p-value for gene1 and gene2 sharing common pathways)
}
\description{
Given the primary gene of interest and the interactor gene list, it calculates the probability of the primary gene and the interactor gene to share common pathways
}
\examples{
gene <- findInteractor()
genepath <- calculatePathway("BRCA2",gene)
}
