% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiscoverSL.r
\name{calculateRNAiScore}
\alias{calculateRNAiScore}
\title{Calculate conditional gene essentiality from RNAi screening}
\usage{
calculateRNAiScore(gene1, gene2)
}
\arguments{
\item{gene1}{character. gene symbol for the primary gene. e.g. "BRCA2"}

\item{gene2}{character vector. It can be either the output gene list from findInteractor, or a user-supplied gene list. list of gene symbols for the interactor gene. e.g. c("PARP1","ATM","NF1","EGFR")}
}
\value{
A data frame with following fields: Gene2(interactor gene), Gene1(primary gene), PvalRNAi(p-value for gene2 having less RNAi score in cells where gene1 is mutated compared to cells where gene1 is not mutated)
}
\description{
Given the primary gene of interest and the interactor gene list, it calculates the difference in RNAi score for targeting the interactor gene in cases where the primary gene is mutated vs not mutated
}
\examples{
dg <- calculateDiffExp("BRCA2","BRCA")
gene <- as.character(dfgene$Gene2)
cg <- calculateExpCorrelation("BRCA2",gene,"BRCA")
gene <- as.character(cg$Gene2)
mutdata <- calculateMutex("BRCA2",gene,"BRCA")
genepath <- calculatePathway("BRCA2",gene)
sl <- predictSL(dg,cg,mutdata,genepath)
gene <- as.character(sl$Gene2)
rnaiscore <- calculateRNAiScore("BRCA2",gene)
}
