% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiscoverSL.r
\name{calculateTTestAmp}
\alias{calculateTTestAmp}
\title{Calculate relative targetability by checking amplification difference}
\usage{
calculateTTestAmp(gene1, gene2, cancer)
}
\arguments{
\item{gene1}{character. gene symbol for the primary gene. e.g. "BRCA2"}

\item{gene2}{character vector. It can be either the output gene list from findInteractor, or a user-supplied gene list. list of gene symbols for the interactor gene. e.g. c("PARP1","ATM","NF1","EGFR")}

\item{cancer}{character. cancer type code. e.g. "BRCA"}
}
\value{
A data frame with following fields: Gene2(interactor gene), Gene1(primary gene), TTestMutAmp(p-value for gene2 having higher amplification of gene2 in cases where gene1 is mutated compared to cases where gene1 is not mutated)
}
\description{
Given the primary gene of interest and the cancer type, it calculates the difference in amplification status of the interactor genes from the supplied gene list (output from findIteractor) in cases where the primary gene is mutated vs not mutated for the cancer type
}
\examples{
gene <- findInteractor()
ampdata <- calculateTTestAmp("BRCA2",gene,"BRCA")
}
