% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiscoverSL.r
\name{customDataDiffExp}
\alias{customDataDiffExp}
\title{Calculate Mutation associated differential expression for user provided mutation and gene expression data}
\usage{
customDataDiffExp(gene1, mymutdata, myexpression)
}
\arguments{
\item{gene1}{character. gene symbol for the primary gene. e.g. "BRCA2"}

\item{mymutdata}{data frame. gene mutation data in tumor samples in the three column format: (gene_symbol, tumor_sample_id, variant_classification).}

\item{myexpression}{data frame. gene expression data in tumor samples in a matrix format. The rows in the matrix will correspond to genes and the columns will correspond to tumor samples.}
}
\value{
A data frame with following fields: Gene2(interactor gene), Gene1(primary gene), logFC(log fold change of expression of gene2 in presence of mutation in gene1), PValue(p-value for differential expression of gene2 in presence of mutation in gene1)
}
\description{
Given the primary gene of interest and the cancer type, it calculates the differential expression of the gene list (output from findIteractor) in samples with or without mutation in the primary gene for the cancer type
}
\examples{
mymutdata
  Hugo_Symbol Tumor_Sample_Barcode Variant_Classification
1       TKTL1      TCGA.AB.2988.03                 Silent
2      GUCA2A      TCGA.AB.2988.03      Missense_Mutation
3      DNMT3A      TCGA.AB.2988.03      Missense_Mutation
4     SLC17A3      TCGA.AB.2988.03      Missense_Mutation
5       LRWD1      TCGA.AB.2988.03      Missense_Mutation
6        BAAT      TCGA.AB.2988.03      Missense_Mutation
myexpression
  Hugo_Symbol TCGA.AB.2803.03 TCGA.AB.2805.03 TCGA.AB.2806.03 TCGA.AB.2807.03 TCGA.AB.2808.03 TCGA.AB.2810.03
1      TKTL1        563.7860        480.5890        415.6588        407.1500        638.6489        350.2327
2      GUCA2A       232.5103        0.6693          103.7556        204.5680        27.4941         36.8666
3      DNMT3A       3169.7531       5251.6734       3499.6817       4340.6157       3394.3441       2516.1456
4      SLC17A3      0.0000          0.0000          0.0000          0.0000          0.0000          0.0000
5      LRWD1        341.5638        167.3360        309.9936        55.6107         197.1720        21.0666
6      BAAT         279.4136        106.4726        218.0204        378.0139        164.0927        155.0899
dg <- customDataDiffExp("BRCA2",mymutdata,myexpression)
}
