% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiscoverSL.r
\name{customDataDrugSensitivity}
\alias{customDataDrugSensitivity}
\title{Calculate relative drug sensitivity (difference in drug sensitivity) for user provided mutation and drug screening data}
\usage{
customDataDrugSensitivity(gene, gene2, celmutdata, celldrug)
}
\arguments{
\item{gene2}{character vector. It can be either the output gene list from findInteractor, or a user-supplied gene list. list of gene symbols for the interactor gene. e.g. c("PARP1","ATM","NF1","EGFR")}

\item{celmutdata}{data frame. gene mutation data in tumor cell lines in the three column format: (gene_symbol, cell_line_id, variant_classification).}

\item{celldrug}{data frame. Drug response data (IC50 values) in tumor cell lines in a matrix format. The rows in the matrix will correspond to genes and the columns will correspond to tumor cell lines.}

\item{gene1}{character. gene symbol for the primary gene. e.g. "BRCA2"}
}
\value{
A data frame with following fields: Gene2(interactor gene), Gene1(primary gene), Drug (Drug name), PvalDrug(p-value for gene2 having less RNAi score in cells where gene1 is mutated compared to cells where gene1 is not mutated)
}
\description{
Given the primary gene of interest and the interactor gene list, it calculates the difference in Drug activity score (IC50) for targeting the interactor gene in cases where the primary gene is mutated vs not mutated
}
\examples{
gene <- findInteractor()
celmutdata
       gene_symbol    case_id mutation
1         AKT3 DMS53_LUNG        0
2         ABI1 DMS53_LUNG        0
3         CDH2 DMS53_LUNG        0
4 LOC100130776 DMS53_LUNG        0
5        HDAC6 DMS53_LUNG        0
6      BCL2L11 DMS53_LUNG        0
celldrug
		Drug	target_gene	COSMIC_906800	COSMIC_910921	COSMIC_687562
1	Erlotinib	EGFR	0.6922463	       2.5381818	2.5602048
2	Sunitinib	PDGFRB	-1.6216855	       -1.7048357   -2.8186215
3	Sunitinib	FLT1	-1.4392042		   -1.2772873   -2.1917983
4	Sunitinib	KIT		-1.3452530		   -1.1004493   -2.9640288
5	Sunitinib	FLT3	-0.7926535		   -0.3372055    0.2775138
6	Sunitinib	CSF1R	-1.4126040		   -0.7056490    1.3978028
drugscore <- calculateDrugSensitivity("BRCA2",gene,celmutdata,celldrug)
}
