% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiscoverSL.r
\name{customDataExpCorrelation}
\alias{customDataExpCorrelation}
\title{Calculate Expression correlation for user provided gene expression data}
\usage{
customDataExpCorrelation(gene1, gene2, myexpression)
}
\arguments{
\item{gene1}{character. gene symbol for the primary gene. e.g. "BRCA2"}

\item{gene2}{character vector. It can be either the output gene list from findInteractor, or a user-supplied gene list. list of gene symbols for the interactor gene. e.g. c("CHEK1","ATM","NF1","EGFR")}

\item{myexpression}{data frame. gene expression data in tumor samples in a matrix format. The rows in the matrix will correspond to genes and the columns will correspond to tumor samples.}
}
\value{
A data frame with following fields: Gene2(interactor gene), Gene1(primary gene), exp.correlation(Pearson's coefficient for expression correlation of gene2 and gene1), correlation.pvalue(p-value for expression correlation of gene2 and gene1)
}
\description{
Given the primary gene of interest and the cancer type, it calculates the expression correlation of the primary gene with the interactor genes from the supplied gene list (output from findIteractor) for the cancer type
}
\examples{
myexpression
  Hugo_Symbol TCGA.AB.2803.03 TCGA.AB.2805.03 TCGA.AB.2806.03 TCGA.AB.2807.03 TCGA.AB.2808.03 TCGA.AB.2810.03
1      TKTL1        563.7860        480.5890        415.6588        407.1500        638.6489        350.2327
2      GUCA2A       232.5103        0.6693          103.7556        204.5680        27.4941         36.8666
3      DNMT3A       3169.7531       5251.6734       3499.6817       4340.6157       3394.3441       2516.1456
4      SLC17A3      0.0000          0.0000          0.0000          0.0000          0.0000          0.0000
5      LRWD1        341.5638        167.3360        309.9936        55.6107         197.1720        21.0666
6      BAAT         279.4136        106.4726        218.0204        378.0139        164.0927        155.0899
gene2 <- findInteractor()
cg <- customDataExpCorrelation("BRCA2",gene2,myexpression)
}
