% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiscoverSL.r
\name{customDataMutex}
\alias{customDataMutex}
\title{Calculate Mutual exclusivity for user provided mutation and CNA data}
\usage{
customDataMutex(gene1, gene2, mycnadata, mymutdata)
}
\arguments{
\item{gene1}{character. gene symbol for the primary gene. e.g. "BRCA2"}

\item{gene2}{character vector. It can be either the output gene list from findInteractor, or a user-supplied gene list. list of gene symbols for the interactor gene. e.g. c("PARP1","ATM","NF1","EGFR")}

\item{mycnadata}{data frame. gene-wise copy number alteration data (GISTIC processed) in tumor samples in a matrix format. The rows in the matrix will correspond to genes and the columns will correspond to tumor samples.}

\item{mymutdata}{data frame. gene mutation data in tumor samples in the three column format: (gene_symbol, tumor_sample_id, variant_classification).}
}
\value{
A data frame with following fields: Gene2(interactor gene), Gene1(primary gene), MutexAmp(p-value for mutual exclusive amplification of gene1 and gene2), MutexDel(p-value for mutual exclusive deletion of gene1 and gene2), MutexMut(p-value for mutual exclusive mutation of gene1 and gene2)
}
\description{
Given the primary gene of interest and the cancer type, it calculates the mutual exclusivity of amplification, deletion and mutation events of the primary gene and the interactor genes from the supplied gene list (output from findIteractor) for the cancer type
}
\examples{
gene <- findInteractor()
mycnadata
  Hugo_Symbol TCGA.3C.AAAU.01 TCGA.3C.AALI.01 TCGA.3C.AALJ.01 TCGA.3C.AALK.01 TCGA.4H.AAAK.01 TCGA.5L.AAT0.01
1 TKTL1               1              -1              -1               0               0               0
2 GUCA2A              1               1               0               0               0               0
3 DNMT3A              1               1               2               0               1               0
4 SLC17A3             1               0               0              -1              -1              -1
5 LRWD1              -1              -1               0               0               0               0
6 BAAT                1               1               0               0               0               0
mymutdata
  Hugo_Symbol Tumor_Sample_Barcode Variant_Classification
1       TKTL1      TCGA.AB.2988.03                 Silent
2      GUCA2A      TCGA.AB.2988.03      Missense_Mutation
3      DNMT3A      TCGA.AB.2988.03      Missense_Mutation
4     SLC17A3      TCGA.AB.2988.03      Missense_Mutation
5       LRWD1      TCGA.AB.2988.03      Missense_Mutation
6        BAAT      TCGA.AB.2988.03      Missense_Mutation
mutdata <- customDataMutex("BRCA2",gene,mycnadata,mymutdata)
}
