% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiscoverSL.r
\name{customDataRNAiScore}
\alias{customDataRNAiScore}
\title{Calculate conditional essentiality (difference in RNAi activity) for user provided mutation and RNAi screening data}
\usage{
customDataRNAiScore(gene1, gene2, celmutdata, cellrnai)
}
\arguments{
\item{gene1}{character. gene symbol for the primary gene. e.g. "BRCA2"}

\item{gene2}{character vector. It can be either the output gene list from findInteractor, or a user-supplied gene list. list of gene symbols for the interactor gene. e.g. c("PARP1","ATM","NF1","EGFR")}

\item{celmutdata}{data frame. gene mutation data in tumor cell lines in the three column format: (gene_symbol, cell_line_id, variant_classification).}

\item{cellrnai}{data frame. RNAi activity data in tumor cell lines in a matrix format. The rows in the matrix will correspond to genes and the columns will correspond to tumor cell lines.}
}
\value{
A data frame with following fields: Gene2(interactor gene), Gene1(primary gene), PvalRNAi(p-value for gene2 having less RNAi score in cells where gene1 is mutated compared to cells where gene1 is not mutated)
}
\description{
Given the primary gene of interest and the interactor gene list, it calculates the difference in RNAi score for targeting the interactor gene in cases where the primary gene is mutated vs not mutated
}
\examples{
celmutdata
       gene_symbol    case_id mutation
1         AKT3 DMS53_LUNG        0
2         ABI1 DMS53_LUNG        0
3         CDH2 DMS53_LUNG        0
4 LOC100130776 DMS53_LUNG        0
5        HDAC6 DMS53_LUNG        0
6      BCL2L11 DMS53_LUNG        0
cellrnai
                          shRNA target_gene X22RV1_PROSTATE X697_HAEMATOPOIETIC_AND_LYMPHOID_TISSUE X786O_KIDNEY
1 GCCCACCAAAGCAAGAAACTT_61E3.4      61E3.4      -1.0261547                              -2.3936993   -0.9911353
2   CAGTGACAGAAGCAGCCATAT_A1BG        A1BG      -1.6216855                              -1.7048357   -2.8186215
3   CCGCCTGTGCTGATGCACCAT_A1BG        A1BG      -1.4392042                              -1.2772873   -2.1917983
4   CTCTTCGAGCTGCACAACATT_A1BG        A1BG      -1.3452530                              -1.1004493   -2.9640288
5   GAGTCCGAATCACTGCTGAAA_A1BG        A1BG      -0.7926535                              -0.3372055    0.2775138
6   CCATGCTGCAAGGAGAGTATA_A1CF        A1CF      -1.4126040                              -0.7056490    1.3978028
gene <- findInteractor()
rnaiscore <- customDataRNAiScore("BRCA2",gene,celmutdata,cellrnai)
}
