% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiscoverSL.r
\name{findInteractor}
\alias{findInteractor}
\title{findInteractor <- function()}
\usage{
findInteractor()
}
\value{
A character vector of gene symbols, of the shortlisted genes
}
\description{
It shortlists genes having mutations in at-least 5 samples in cancers. This gene list is used in the later functions to predict SL partners
}
\examples{
gene <- findInteractor()
}
