% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiscoverSL.r
\name{plotAmplificationDiff}
\alias{plotAmplificationDiff}
\title{Plot relative amplification difference for SL gene}
\usage{
plotAmplificationDiff(gene1, gene2, cancer)
}
\arguments{
\item{gene1}{character. gene symbol for the primary gene. e.g. "BRCA2"}

\item{gene2}{character. gene symbol for the primary gene. e.g. "PARP1"}

\item{cancer}{character. cancer type code. e.g. "BRCA"}
}
\description{
Given the primary gene of interest and the cancer type, it calculates the difference in amplification status of the interactor gene in cases where the primary gene is mutated vs not mutated for the cancer type and shows the difference in a box plot
}
\examples{
plotAmplificationDiff("BRCA1","PARP1","BRCA")
}
