% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiscoverSL.r
\name{plotDrugSensitivity}
\alias{plotDrugSensitivity}
\title{Plot conditional sensitivity to any drugs targeting SL gene}
\usage{
plotDrugSensitivity(gene1, gene2)
}
\arguments{
\item{gene1}{character. gene symbol for the primary gene. e.g. "BRCA2"}

\item{gene2}{character. gene symbol for the interactor gene. e.g. "PARP1"}
}
\description{
Given the primary gene of interest and the interactor gene, it calculates the difference in Drug activity score (IC50) for targeting the interactor gene in cases where the primary gene is mutated vs not mutated and shows the difference in a box plot
}
\examples{
plotDrugSensitivity("BRCA1","PARP1")
}
