% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiscoverSL.r
\name{plotRNAi}
\alias{plotRNAi}
\title{Plot conditional gene essentiality from Achilles RNAi screening}
\usage{
plotRNAi(gene1, gene2)
}
\arguments{
\item{gene1}{character. gene symbol for the primary gene. e.g. "BRCA2"}

\item{gene2}{character. gene symbol for the interactor gene. e.g. "PARP1"}

\item{mcle}{data frame. output from loadMutationDataCCLE}

\item{mrnai}{data frame. output from loadRNAidataAchilles}
}
\description{
Given the primary gene of interest and the interactor gene, it calculates the difference in RNAi score for targeting the interactor gene in cases where the primary gene is mutated vs not mutated and shows the difference in a box plot
}
\examples{
plotRNAi("BRCA1","PARP1")
}
