% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiscoverSL.r
\name{plotSensitivitybyDrug}
\alias{plotSensitivitybyDrug}
\title{Plot conditional sensitivity for a given drug}
\usage{
plotSensitivitybyDrug(gene1, gene2, drug)
}
\arguments{
\item{gene1}{character. gene symbol for the primary gene. e.g. "BRCA2"}

\item{gene2}{character. gene symbol for the interactor gene. e.g. "PARP1"}

\item{drug}{character. drug name. e.g. "Olaparib"}
}
\description{
Given the primary gene of interest, the interactor gene and the drug name targeting the interactor gene, it calculates the difference in Drug activity score (IC50) for targeting the interactor gene in cases where the primary gene is mutated vs not mutated and shows the difference in a box plot
}
\examples{
plotSensitivitybyDrug("BRCA1","PARP1","Olaparib")
}
