% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiscoverSL.r
\name{plotSurvivalCurveSL}
\alias{plotSurvivalCurveSL}
\title{Plot Survival difference for SL co-inactivation}
\usage{
plotSurvivalCurveSL(gene1, gene2, cancer)
}
\arguments{
\item{gene1}{character. gene symbol for the primary gene. e.g. "BRCA2"}

\item{gene2}{character. gene symbol for the interactor gene. e.g. "PARP1"}

\item{cancer}{character. cancer type code. e.g. "BRCA"}
}
\description{
Given the primary gene of interest and the cancer type, it calculates the difference in patient disease-free survival depending on the co-inactivation of the primary gene and the interactor gene for the cancer type
}
\examples{
plotSurvivalCurveSL("BRCA1","PARP1","BRCA")
}
