% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiscoverSL.r
\name{predictSL}
\alias{predictSL}
\title{Predict synthetic lethality}
\usage{
predictSL(dg, cg, mutdata, genepath)
}
\arguments{
\item{dg}{data frame. Output from calculateDiffExp}

\item{cg}{data frame. Output from calculateExpCorrelation}

\item{mutdata}{data frame. Output from calculateMutex}

\item{genepath}{data frame. Output from calculatePathway}
}
\value{
A data frame with following fields: Gene2(interactor gene), Gene1(primary gene), Diffexp.Pvalue(p-value for differential expression of gene2 in presence of mutation in gene1), correlation.pvalue(p-value for expression correlation of gene2 and gene1), MutexMut(p-value for mutual exclusive mutation of gene1 and gene2), PvalPathway(p-value for gene1 and gene2 sharing common pathways), MetaFDR (FDR corrected combined P-value from four parameters), SyntheticLethalScore(Predicted synthetic lethality score between gene1 and gene2)
}
\description{
Given the primary gene of interest and the cancer type, it calculates the Synthetic lethality score for all interactor genes from the supplied gene list (output from findIteractor) for the cancer type, using the four parameters calculated from functions calculateDiffExp, calculateExpCorrelation, calculateMutex and calculatePathway
}
\examples{
dg <- calculateDiffExp("BRCA2","BRCA")
gene <- findInteractor()
cg <- calculateExpCorrelation("BRCA2",gene,"BRCA")
gene <- as.character(cg$Gene2)
mutdata <- calculateMutex("BRCA2",gene,"BRCA")
genepath <- calculatePathway("BRCA2",gene)
sl <- predictSL(dg,cg,mutdata,genepath)
}
