% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiscoverSL.r
\name{summarizeSLParams}
\alias{summarizeSLParams}
\title{Summarize potential synthetic lethals by all parameters}
\usage{
summarizeSLParams(sl, rnaiscore, drugscore, ampdata)
}
\arguments{
\item{sl}{data frame. Output from predictSL}

\item{rnaiscore}{data frame. Output from calculateRNAiScore}

\item{drugscore}{data frame. Output from calculateDrugSensitivity}

\item{ampdata}{data frame. Output from calculateTTestAmp}
}
\value{
A data frame with following fields: Gene1(primary gene), Gene2(interactor gene), Diffexp.Pvalue(p-value for differential expression of gene2 in presence of mutation in gene1), correlation.pvalue(p-value for expression correlation of gene2 and gene1), MutexMut(p-value for mutual exclusive mutation of gene1 and gene2), PvalPathway(p-value for gene1 and gene2 sharing common pathways), MetaFDR (FDR corrected combined P-value from four parameters), SyntheticLethalScore(Predicted synthetic lethality score between gene1 and gene2), PvalRNAi (P-value for conditional essentiality from RNAi screens), Drug (Drug name), PvalDrug (P-value for conditional drugsensitivity), TTestMutAmp (p-value for conditional targetability)
}
\description{
Given the prior calculated parameters SyntheticLethalScore, RNAiscore,DrugSensitivity and TargetAmplification, it summarizes all these parameters in a single table for all interactor genes
}
\examples{
dg <- calculateDiffExp("BRCA2","BRCA")
gene <- findInteractor()
cg <- calculateExpCorrelation("BRCA2",gene,"BRCA")
gene <- as.character(cg$Gene2)
mutdata <- calculateMutex("BRCA2",gene,"BRCA")
genepath <- calculatePathway("BRCA2",gene)
sl <- predictSL(dg,cg,mutdata,genepath)
rnaiscore <- calculateRNAiScore(gene1,gene)
drugscore <- calculateDrugSensitivity(gene1,gene)
ampdata <- calculateTTestAmp(gene1,gene,cancer)
finalist <- summarizeSLParams(sl,rnaiscore,drugscore,ampdata)
}
